// types.ts
export interface StudentData {
  nama: string;
  nim: string;
  prodi: string;
  tempatLahir: string;
  tanggalLahir: string;
  alamat1: string;
  alamat2: string;
  golDarah: string;
  phone: string;
  email: string;
  masaBerlaku: string;
  logo: string | null;
  foto: string | null;
  stempel: string | null;
  tandaTangan: string | null;
}